requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-WTF>=1.2
Flask-Limiter>=3.8
python-dotenv>=1.0
psycopg[binary]>=3.2
gunicorn>=22.0
pytest>=8.0
